assignments/week05/athome/requirements.txt
Beaker==1.6.4
SQLAlchemy==0.8.0b2
bottle==0.11.6
bottle-cork==0.7a
bottle-pystache==0.0.3
bottle-sqlalchemy==0.3.1
pycrypto==2.6
wsgiref==0.1.2
